Move scaling helpers to typedef file.
