+++ src/structures/typedefs.h
@@ -149,6 +149,20 @@ enum OperatorName {
   MAX
 };
 
+namespace utils {
+inline Duration scale_from_user_duration(UserDuration d) {
+  return DURATION_FACTOR * static_cast<Duration>(d);
+}
+
+inline UserDuration scale_to_user_duration(Duration d) {
+  return static_cast<UserDuration>(d / DURATION_FACTOR);
+}
+
+inline UserCost scale_to_user_cost(Cost d) {
+  return static_cast<UserCost>(d / DURATION_FACTOR);
+}
+} // namespace utils
+
 #ifdef LOG_LS_OPERATORS
 namespace ls {
 struct OperatorStats {
