+++ src/utils/helpers.h
@@ -45,18 +45,6 @@ inline UserCost add_without_overflow(UserCost a, UserCost b) {
   return a + b;
 }
 
-inline Duration scale_from_user_duration(UserDuration d) {
-  return DURATION_FACTOR * static_cast<Duration>(d);
-}
-
-inline UserDuration scale_to_user_duration(Duration d) {
-  return static_cast<UserDuration>(d / DURATION_FACTOR);
-}
-
-inline UserCost scale_to_user_cost(Cost d) {
-  return static_cast<UserCost>(d / DURATION_FACTOR);
-}
-
 inline INIT get_init(const std::string& s) {
   if (s == "NONE") {
     return INIT::NONE;
